app/requirements_ubuntu.txt
absl-py==2.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
backports.tarfile==1.2.0
bcrypt==4.1.2
bidict==0.23.1
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
gotrue==2.7.0
greenlet==3.0.3
grpcio==1.66.0
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
install-jdk==1.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
keras==2.15.0
keyring==25.3.0
konlpy==0.6.0
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
nh3==0.2.18
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.20.0
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.2.2
postgrest==0.16.11
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
PyYAML==6.0.2
readme_renderer==44.0
realtime==1.0.6
redis==5.0.8
reflex==0.5.10
reflex-calendar==0.0.4
reflex-chakra==0.5.10
reflex-hosting-cli==0.1.13
reflex-wordcloud==0.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.2.2
scipy==1.14.1
SecretStorage==3.3.3
sentence-transformers==2.7.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.37.2
starlette-admin==0.14.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.7
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.40.1
triton==2.2.0
twine==5.1.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
websockets==12.0
Werkzeug==3.0.4
wrapt==1.14.1
wsproto==1.2.0
zipp==3.20.1
